ref(api): Move get_date_range_from_params into utils
